requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
email-validator==2.2.0
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
bcrypt==3.2.0
pydantic==2.12.0
pydantic_core==2.41.1
pydantic-settings==2.5.2
python-dotenv==1.1.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.0
python-jose[cryptography]==3.3.0

